Casts (and the like, mostly to ints) to remove compiler warnings on Lobo.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@2271 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/columns/HyPerCol.cpp
@@ -67,8 +67,8 @@ HyPerCol::HyPerCol(const char * name, int argc, char * argv[])
    }
 
    if (status) {
-      imageLoc.nx = params->value(name, "nx");
-      imageLoc.ny = params->value(name, "ny");
+      imageLoc.nx = (int) params->value(name, "nx");
+      imageLoc.ny = (int) params->value(name, "ny");
       imageLoc.nxGlobal = imageLoc.nx * icComm->numCommColumns();
       imageLoc.nyGlobal = imageLoc.ny * icComm->numCommRows();
       imageLoc.kx0 = imageLoc.nx * icComm->commColumn();
